+++ gui/next/src/lib/api/record.js
@@ -5,18 +5,14 @@
 
 // imports
 // ------------------------------------------------------------------------
-import JSON5 from 'json5'
-
-import { tryParseJSON } from '$lib/tryParseJSON.js';
 import { graphql } from '$lib/api/graphql';
 import { state } from '$lib/state';
 
 
 
-
-// purpose:		maps GraphQL types to corresponding GraphQL type query
+// purpose:   maps column types to corresponding property type used in GraphQL string
 // ------------------------------------------------------------------------
-const typeMap = {
+const columnTypeToPropertyType = {
   array: 'value_array',
   boolean: 'value_boolean',
   date: 'value',
@@ -29,98 +25,201 @@ const typeMap = {
   json: 'value_json'
 };
 
-
-// purpose:		builds properties string from FormData to pass to GraphQL
-// arguments:	FormData properties serialized to object (object) - Object.fromEntries(<FormData>.entries())
-// returns:		GraphQL string with properties (string)
+// purpose:		maps column types to corresponding variable type used in GraphQL variables definition
 // ------------------------------------------------------------------------
-const getPropertiesString = (props) => {
-
-  let output = {};
-
-  Object.keys(props)
-  .forEach((prop) => {
-
-    let property = prop.slice(0, prop.indexOf('['));
-    let key = prop.slice(prop.indexOf('[')+1, prop.indexOf(']'));
-
-    (output[property] ??= {})[key] = props[prop];
-  });
-
-  let string = '';
-
-  Object.keys(output).forEach(key => {
-    // we are showing escaped JSONs as indented tree, so we need to remove new line signs and indentation when saved
-    if(output[key].type === 'string' && tryParseJSON(output[key].value)){
-      output[key].value = JSON.stringify(JSON.parse(output[key].value));
-    }
-
-    if(output[key].type === 'string' || output[key].type === 'text' || output[key].type === 'date' || output[key].type === 'time' || output[key].type === 'datetime'){
-      output[key].value = JSON.stringify(output[key].value);
-    }
-
-    // so if this is a true JSON we expect to send it to GraphQL api like a JS-object but in a string
-    // that's why it needs to be parsed from editing JSON {"something": "something"} to {something: "something"}
-    if(output[key].type === 'json'){
-      output[key].value = JSON5.stringify(JSON.parse(output[key].value), { quote: '"' });
-    }
-
-    if(!output[key].value){
-      output[key].value = null;
-    }
-
-    // don't allow editing the upload property but allow any other
-    if(output[key].type !== 'upload'){
-      string += `{ name: "${key}", ${typeMap[output[key].type]}: ${output[key].value} }`;
-    }
-  });
-
-  return string;
+const columnTypeToVariableType = {
+  string: 'String',
+  integer: 'Int',
+  float: 'Float',
+  boolean: 'Boolean',
+  array: '[String!]',
+  json: 'JSONPayload',
+  range: 'RangeFilter'
 };
 
 
-// purpose:		build the filters string to pass to GraphQL
-// arguments:	list of attributes to filter the data with (array of objects) that icludes:
-//				    attribute_type, property, operation, value
-// returns:		GraphQL string with filtering properties (string)
-// ------------------------------------------------------------------------
-const getFiltersString = (filters) => {
-  let filtersString = '';
 
-  const operation_types = {
+// purpose:		build the strings and objects needed to pass with GraphQL request to filter the properties
+// arguments:	list of properties to filter the data with (array of objects) that icludes:
+//				    attribute_type, property, operation, value
+// returns:		GraphQL variables definitions, e.g. '($variable_name: String. $another_variable: Int)' (string)
+//            variables with their values to pass with the query, e.g. { variable_name: 'Variable value', another_variable: 5 } (object)
+//            filters used to filter properties in GraphQL requests, e.g. 'properties: [name: "variable_name", contains: $variable_value]' (string)
+// ------------------------------------------------------------------------
+const buildQueryIngredients = (filters = []) => {
+  // graphql variables definition for the query (string)
+  let variablesDefinition = '';
+  // variables passed with the query (object)
+  let variables = {};
+  // filters passed with the query (string)
+  let propertiesFilter = '';
+  // list of data types for variables in each operations, unmentioned are considered 'string' (object)
+  const operationsForType = {
     string: ['array_contains', 'not_array_contains', 'contains', 'ends_with', 'not_contains', 'not_ends_with', 'not_starts_with', 'not_value', 'starts_with', 'value'],
-    range: ['range']
+    int: ['value_int', 'not_value_int'],
+    float: ['not_value_float', 'value_float'],
+    bool: ['exists', 'not_value_boolean', 'value_boolean'],
+    range: ['range'],
+    array: ['value_array', 'not_value_array', 'value_in', 'not_value_in', 'array_overlaps', 'not_array_overlaps']
   };
 
-  filters.forEach(filter => {
+  // build the data for each applied filter
+  for(const filter of filters){
 
-    let parsedValue = '';
-
-    if(operation_types.string.includes(filter.operation)){
-      parsedValue = `"${filter.value}"`;
+    // if there is no value, don't output filter string for that filter which effectively clears it
+    if(!filter.minFilterValue && !filter.maxFilterValue && !filter.value){
+      break;
     }
-    else if(operation_types.range.includes(filter.operation)){
-      parsedValue = `{ ${filter.minFilter}: "${filter.minFilterValue}", ${filter.maxFilter}: "${filter.maxFilterValue}" }`;
+
+    // storing the type for current filter
+    let filterType = '';
+    // we are getting each filter value as a string, so it needs to be parsed for graphql request
+    let parsedFilterValue = '';
+
+    if(operationsForType.int.includes(filter.operation)){
+      filterType = 'integer';
+      parsedFilterValue = parseInt(filter.value);
+    }
+    else if (operationsForType.float.includes(filter.operation)){
+      filterType = 'float';
+      parsedFilterValue = parseFloat(filter.value);
+    }
+    else if (operationsForType.bool.includes(filter.operation)){
+      filterType = 'boolean';
+      parsedFilterValue = filter.value === 'true' ? true : false;
+    }
+    else if(operationsForType.range.includes(filter.operation)){
+      filterType = 'range';
+      parsedFilterValue = {};
+      parsedFilterValue[filter.minFilter] = filter.minFilterValue;
+      parsedFilterValue[filter.maxFilter] = filter.maxFilterValue;
+    }
+    else if(operationsForType.array.includes(filter.operation)){
+      filterType = 'array';
+      parsedFilterValue = JSON.parse(filter.value);
     }
     else {
-      parsedValue = filter.value;
+      filterType = 'string';
+      parsedFilterValue = filter.value;
     }
 
-    // filtering by id is not done in this string
-    if(filter.name !== 'id' && parsedValue){
-      filtersString += `{
+    // skipping the ID as it is not filtered as a property
+    if(filter.name !== 'id'){
+      // add current filter variable to variables definition string
+      variablesDefinition += `, $${filter.name}: ${columnTypeToVariableType[filterType]}`;
+
+      // add the current filter to the variables object passed with the request (corresponding with variables definition)
+      variables[filter.name] = parsedFilterValue;
+
+      // add current filter to properties filters string passed in GraphQL request
+      propertiesFilter += `{
         name: "${filter.name}",
-        ${filter.operation}: ${parsedValue}
+        ${filter.operation}: $${filter.name}
       }`;
     }
 
-  });
+  };
 
-  filtersString = `
-    properties: [${filtersString}]
+  // build the final variables definition string with all the needed variables and their types
+  if(variablesDefinition.length){
+    variablesDefinition = variablesDefinition.slice(2); // remove first comma ', '
+    variablesDefinition = `(${variablesDefinition})`; // add brackets to definition string
+  }
+
+  // build final string for filtering the properties
+  propertiesFilter = `
+    properties: [${propertiesFilter}]
   `;
 
-  return filtersString;
+  return { variablesDefinition, variables, propertiesFilter };
+};
+
+
+
+// purpose:		builds all the needed data to build and trigger GraphQL mutation
+// arguments: FormData object with the column, type and value (column[value], column[type])
+// returns:   variablesDefinition (string) - GraphQL variables definitions $variable: Type
+//            variables (object) - values for defined variables passed to GraphQL API
+//            properties (string) - list of GraphQL properties for given variables to use inside "properties: []"
+// ------------------------------------------------------------------------
+const buildMutationIngredients = (formData) => {
+
+  // entries from the form (object)
+  const formEntries = formData.entries();
+  // graphql variables definition for the query (string)
+  let variablesDefinition = '';
+  // variables passed with the query (object)
+  let variables = {};
+  // properties list with variables as their value passed with the query (string)
+  let properties = '';
+  // helper object that will store the column name with all it's corresponding propertiest needed to pass to graphql (object)
+  let columns = {};
+
+  // values we are getting from FormData are strings, this will parse them depending on the column type we are taking them from
+  function parseValue(type, value){
+    if(value === ''){
+      return null;
+    }
+
+    if(type === 'integer'){
+      return parseInt(value);
+    }
+
+    if(type === 'float'){
+      return parseFloat(value);
+    }
+
+    if(type === 'boolean'){
+      if(value === 'true'){
+        return true;
+      } else {
+        return false;
+      }
+    }
+
+    if(type === 'array'){
+      return JSON.parse(value);
+    }
+
+    if(type === 'json'){
+      return JSON.parse(value);
+    }
+
+    return value;
+  };
+
+  // parse FormData entries to an `columns` object
+  for(const [entry, value] of formEntries){
+    // do it only for FormData entries related to columns
+    if(entry.indexOf('[') >= 0){
+      // get the column name from FormData
+      let column = entry.slice(0, entry.indexOf('['));
+      // get the column properties names like the 'type' and 'value'
+      let property = entry.slice(entry.indexOf('[')+1, entry.indexOf(']'));
+
+      // build an object like: column_name = { type: 'type', value: 'value' }
+      (columns[column] ??= {})[property] = value;
+    }
+  };
+
+  // for each edited column build the needed strings and add the value to `variables`
+  for(const column in columns){
+    variablesDefinition += `, $${column}: ${columnTypeToVariableType[columns[column].type]}`;
+
+    variables[column] = parseValue(columns[column].type, columns[column].value);
+
+    properties += `{ name: "${column}", ${columnTypeToPropertyType[columns[column].type]}: $${column} }`;
+  }
+
+
+  // build the final variables definition string with all the needed variables and their types
+  if(variablesDefinition.length){
+    variablesDefinition = variablesDefinition.slice(2); // remove first comma ', '
+    variablesDefinition = `(${variablesDefinition})`; // add brackets to definition string
+  }
+
+  return { variablesDefinition, variables, properties }
+
 };
 
 
@@ -135,19 +234,18 @@ const record = {
   // returns:		array of records as they appear in the database (array)
   // ------------------------------------------------------------------------
   get: (args) => {
+
     const defaults = {
       deleted: false,
       filters: {
         page: 1
       }
     };
-
     const params = {...defaults, ...args};
 
     const tableFilter = params.table ? `table_id: { value: ${params.table} }` : '';
 
     const idFilterIndex = params.filters?.attributes?.findIndex(attribute => attribute.name === 'id');
-
     let idFilter = '';
     if(idFilterIndex >= 0 && params.filters.attributes[idFilterIndex].value){
       idFilter = `id: { ${params.filters.attributes[idFilterIndex].operation}: ${params.filters.attributes[idFilterIndex].value} }`;
@@ -166,10 +264,10 @@ const record = {
 
     const deletedFilter = params.deleted === 'true' ? `deleted_at: { exists: true }` : '';
 
-    const filters = params.filters?.attributes ? getFiltersString(params.filters.attributes) : '';
+    const propertiesFilterData = buildQueryIngredients(params.filters?.attributes);
 
     const query = `
-      query {
+      query${propertiesFilterData.variablesDefinition} {
         records(
           page: ${params.filters.page}
           per_page: 20,
@@ -178,7 +276,7 @@ const record = {
             ${tableFilter}
             ${idFilter}
             ${deletedFilter}
-            ${filters}
+            ${propertiesFilterData.propertiesFilter}
           }
         ) {
           current_page
@@ -193,33 +291,30 @@ const record = {
         }
       }`;
 
-    return graphql({ query }).then(data => { state.data('records', data.records) });
+    return graphql({ query, variables: propertiesFilterData.variables }).then(data => { state.data('records', data.records) });
   },
 
 
   // purpose:		creates new record in the database
-  // arguments:	(object)
-  //				    tableName (string) - name of the table that you are adding the record in
-  //				    properties (FormData) - key-value pairs for the record
+  // arguments:
   // returns:		id of the newly created record (int)
   // ------------------------------------------------------------------------
   create: (args) => {
-    let properties = Object.fromEntries(args.properties.entries());
-    const table = properties.tableName;
-    delete properties.tableName;
-    properties = getPropertiesString(properties);
+    const formDataEntries = Object.fromEntries(args.properties.entries());
+    const table = formDataEntries.tableName;
+    const ingredients = buildMutationIngredients(args.properties);
 
     const query = `
-      mutation {
+      mutation${ingredients.variablesDefinition} {
         record_create(record: {
           table: "${table}",
-          properties: [${properties}]
+          properties: [${ingredients.properties}]
         }) {
           id
         }
       }`;
 
-    return graphql({ query });
+    return graphql({ query, variables: ingredients.variables });
   },
 
 
@@ -231,27 +326,25 @@ const record = {
   // returns:		id of the edited record (int)
   // ------------------------------------------------------------------------
   edit: (args) => {
-    let properties = Object.fromEntries(args.properties.entries());
-    const table = properties.tableName;
-    delete properties.tableName;
-    const id = properties.recordId;
-    delete properties.recordId;
-    properties = getPropertiesString(properties);
+    let formDataEntries = Object.fromEntries(args.properties.entries());
+    const table = formDataEntries.tableName;
+    const id = formDataEntries.recordId;
+    const ingredients = buildMutationIngredients(args.properties);
 
     const query = `
-      mutation {
+      mutation${ingredients.variablesDefinition} {
         record_update(
           id: ${id},
           record: {
             table: "${table}"
-            properties: [${properties}]
+            properties: [${ingredients.properties}]
           }
         ) {
           id
         }
       }`;
 
-    return graphql({ query });
+    return graphql({ query, variables: ingredients.variables });
   },